+++ app/src/main/java/de/baumann/sieben/helper/UserSettingsActivity.java
@@ -2,9 +2,12 @@ package de.baumann.sieben.helper;
 
 import android.app.AlertDialog;
 import android.content.DialogInterface;
+import android.content.Intent;
+import android.net.Uri;
 import android.os.Bundle;
 import android.preference.Preference;
 import android.preference.PreferenceFragment;
+import android.preference.PreferenceManager;
 import android.support.v7.app.AppCompatActivity;
 import android.text.Html;
 import android.text.SpannableString;
@@ -29,6 +32,9 @@ public class UserSettingsActivity extends AppCompatActivity {
         getFragmentManager().beginTransaction()
                             .replace(android.R.id.content, new SettingsFragment())
                             .commit();
+
+        PreferenceManager.setDefaultValues(this, R.xml.user_settings, false);
+        PreferenceManager.setDefaultValues(this, R.xml.user_settings_exercises, false);
     }
 
     @Override
@@ -42,33 +48,11 @@ public class UserSettingsActivity extends AppCompatActivity {
         private void addChangelogListener() {
             Preference reset = findPreference("changelog");
 
-            reset.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener()
-            {
-                public boolean onPreferenceClick(Preference pref)
-                {
-
-                    SpannableString s;
-
-                    if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
-                        s = new SpannableString(Html.fromHtml(getString(R.string.changelog_text),Html.FROM_HTML_MODE_LEGACY));
-                    } else {
-                        s = new SpannableString(Html.fromHtml(getString(R.string.changelog_text)));
-                    }
-
-                    Linkify.addLinks(s, Linkify.WEB_URLS);
-
-                    final AlertDialog d = new AlertDialog.Builder(getActivity())
-                            .setTitle(R.string.action_changelog)
-                            .setMessage( s )
-                            .setPositiveButton(getString(R.string.about_yes),
-                                    new DialogInterface.OnClickListener() {
-                                        public void onClick(DialogInterface dialog, int id) {
-                                            dialog.cancel();
-                                        }
-                                    }).show();
-                    d.show();
-                    ((TextView)d.findViewById(android.R.id.message)).setMovementMethod(LinkMovementMethod.getInstance());
-
+            reset.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+                public boolean onPreferenceClick(Preference pref) {
+                    Uri uri = Uri.parse("https://github.com/scoute-dich/Sieben/blob/master/CHANGELOG.md"); // missing 'http://' will cause crashed
+                    Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+                    startActivity(intent);
                     return true;
                 }
             });
@@ -93,6 +77,7 @@ public class UserSettingsActivity extends AppCompatActivity {
                     Linkify.addLinks(s, Linkify.WEB_URLS);
 
                     final AlertDialog d = new AlertDialog.Builder(getActivity())
+                            .setTitle(R.string.action_help)
                             .setMessage(s)
                             .setPositiveButton(getString(R.string.about_yes),
                                     new DialogInterface.OnClickListener() {
@@ -143,6 +128,35 @@ public class UserSettingsActivity extends AppCompatActivity {
             });
         }
 
+        private void addDonateListListener() {
+
+            Preference reset = findPreference("donate");
+            reset.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+                public boolean onPreferenceClick(Preference pref) {
+                    Uri uri = Uri.parse("https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=NP6TGYDYP9SHY"); // missing 'http://' will cause crashed
+                    Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+                    startActivity(intent);
+                    return true;
+                }
+            });
+        }
+
+        private void add_exerciseChooseListener() {
+
+            Preference reset = findPreference("exercises");
+
+            reset.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+                public boolean onPreferenceClick(Preference pref) {
+
+                    Intent intent_in = new Intent(getActivity(), UserSettingsActivity_Exercises.class);
+                    startActivity(intent_in);
+                    getActivity().overridePendingTransition(0, 0);
+
+                    return true;
+                }
+            });
+        }
+
 
         @Override
         public void onCreate(Bundle savedInstanceState) {
@@ -152,6 +166,8 @@ public class UserSettingsActivity extends AppCompatActivity {
             addHelpListener();
             addLicenseListener();
             addChangelogListener();
+            addDonateListListener();
+            add_exerciseChooseListener();
         }
     }
 }