+++ app/src/main/java/de/baumann/sieben/helper/UserSettingsActivity_Exercises.java
@@ -1,0 +1,85 @@
+/*
+    This file is part of the Browser WebApp.
+
+    Browser WebApp is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    Browser WebApp is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with the Browser webview app.
+
+    If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package de.baumann.sieben.helper;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.preference.PreferenceFragment;
+import android.preference.PreferenceManager;
+import android.support.v7.app.AppCompatActivity;
+import android.view.MenuItem;
+
+import de.baumann.sieben.R;
+
+
+public class UserSettingsActivity_Exercises extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        android.support.v7.app.ActionBar actionBar = getSupportActionBar();
+        assert actionBar != null;
+        actionBar.setDisplayHomeAsUpEnabled(true);
+        setTitle(R.string.action_exercises);
+
+        getFragmentManager().beginTransaction()
+                .replace(android.R.id.content, new SettingsFragment())
+                .commit();
+
+        PreferenceManager.setDefaultValues(this, R.xml.user_settings, false);
+        PreferenceManager.setDefaultValues(this, R.xml.user_settings_exercises, false);
+    }
+
+    public static class SettingsFragment extends PreferenceFragment {
+
+        @Override
+        public void onCreate(Bundle savedInstanceState) {
+            super.onCreate(savedInstanceState);
+
+            addPreferencesFromResource(R.xml.user_settings_exercises);
+        }
+    }
+
+    @Override
+    public void onBackPressed() {
+        Intent intent_in = new Intent(UserSettingsActivity_Exercises.this, UserSettingsActivity.class);
+        startActivity(intent_in);
+        overridePendingTransition(0, 0);
+        finish();
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+        int id = item.getItemId();
+
+        if (id == android.R.id.home) {
+            Intent intent_in = new Intent(UserSettingsActivity_Exercises.this, UserSettingsActivity.class);
+            startActivity(intent_in);
+            overridePendingTransition(0, 0);
+            finish();
+
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+}