+++ app/src/main/java/de/baumann/sieben/workouts/MainActivity1.java
@@ -54,6 +54,7 @@ public class MainActivity1 extends AppCompatActivity {
         setContentView(R.layout.activity_main);
 
         PreferenceManager.setDefaultValues(this, R.xml.user_settings, false);
+        PreferenceManager.setDefaultValues(this, R.xml.user_settings_exercises, false);
         SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(MainActivity1.this);
         final String dur = sharedPref.getString("duration", "30");
         duration = Integer.parseInt(dur);